LUS-642 lprocfs: preallocate all lprocfs_stats
Allocating the stats as-needed leads to fragmentation.

lprocfs_alloc_stats() preallocates only for IRQ_SAFE stats.  It does
this so that it doesn't have to try to allocate memory in an interrupt
handler, not for fragmentation reasons.

The test for IRQ_SAFE is #ifdef'd, so that we preallocate space for
all the stats.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: Iaed09f6763fe53e273672829bece417c9370a15e
Reviewed-on: http://es-gerrit.dev.cray.com:8080/17263
Reviewed-by: Justin Miller <jmiller@cray.com>
Reviewed-by: Alexander Boyko <c17825@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>