Set ping interval to 25 seconds
This is a temporary fix to help us recover more quickly after
a quiesce. Many sites currently use an obd_timeout of 300 seconds.
This results in a ping interval of 75 seconds. Experimentation shows
that the first reconnect after a quiesce almost always fails. If
the servers have a failover partner then the second connection attempt
will go to the failover partner which will also fail. The third
connection attempt typically succeeds. With this mod, instead of
3*75 seconds to get reconnected we should only need 3*25 (+ net latency).
Bugs Affected: 812663
Bug Action: UPDATE
Test Description:
Passed sanity in sdmvm cluster. Originally tested on galaxy with a
combination of blade failure and failover/failback scenarios.

svn-id: lustre:20841