LU-5580 ptlrpc: policy switch directly in tbf
Former version can't change nid to jobid(or reversely) directly in
tbf, it need to change to FIFO(or any other policy) first. This patch
fix this problem. It save the policy argument(like "nid" in tbf),
and if the policy start with a new argument(like "jobid"), it will
stop the former policy and start a new one.

This is also adapt to any other nrs policies if they have arguments.

This patch is backported from the following one:
Lustre-commit: e7ab554c1ca887e1a3fa9da5250b2debb4eee2d6
Lustre-Change: http://review.whamcloud.com/11749

Change-Id: I35e3aa377b1a727a8316d75ed4b0112517ce2c36
Signed-off-by: Wu Libin <lwu@ddn.com>
Signed-off-by: Emoly Liu <emoly.liu@intel.com>
Reviewed-on: http://review.whamcloud.com/16156
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>