cray-lustre-default install fix
This is a temporary fix to this issue with the imps image creation

	Return code: 1
	Response: Error: Package: robinhood-2.99-14.19.x86_64 (var_opt_cray_repos_common_pinehurst_trunk_centos_7.0_x86-64)
           Requires: liblustreapi.so()(64bit)
 You could try using --skip-broken to work around the problem
 You could try running: rpm -Va --nofiles --nodigest

This is not the correct way we should go about doing this but the solution
is being tracked in OBS-299. With the fix provided by this mod the imps
image builds without issue. This will be removed when we figure out what
the actual problem is.

Test Description:
I built the image recipe on ph3 and verified that lustre
installed without issue.

[root@ph3-m000 repos]# rpm -qp --provides cray-lustre-default-2.7_3.10.86_9.1-1.0000.20948.0.0.x86_64.rpm
cray-lustre-default = 2.7_3.10.86_9.1-1.0000.20948.0.0
cray-lustre-default(x86-64) = 2.7_3.10.86_9.1-1.0000.20948.0.0
liblustreapi.so()(64bit)

It should also be noted tha liblustreapi.so is contained in the rpm. From
rpm -qlip: /usr/lib64/liblustreapi.so

None

svn-id: lustre:21000