LU-7015 ofd: Fix wanted grant calculation
When clients sends a "we want this much grant number", the value
is a total amount of space client wants including whatever it has.
So in order to properly reconcile this on the server we need
to substract what the client alreay has to see how much we grant back.
This will still overshoot some due to multiple RPCs in flight,
but at least it should not be as bad anymore.

This patch is back-ported from the following one:
Lustre-commit: 091988499717c729f8870b331ab3774b249d5818
Lustre-change: http://review.whamcloud.com/16216

Signed-off-by: Oleg Drokin <oleg.drokin@intel.com>
Change-Id: I558b320d2473335b74f59ebb44be735fdaaa8d3a
Reviewed-on: http://review.whamcloud.com/16354
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>