Release ldlm_namespace_lock before spawning a fixed number of threads in ldlm_drop_caches
Port for Cray-b2_7.

This mod restructures the drop_caches code so that it releases the lock on the
namespace list before spawning threads.  It also spawns a fixed number of
threads which have access to a work queue of namespaces to be cleared.

The new algorithm is roughly as follows.  Allocate a work queue which can hold
a work item for each namespace.  Lock the namespace list.  Iterate through the
list of active namespaces, adding a work item for each one to the work queue.
Release the lock on the namespace list.  For a fixed number of threads, set up
a thread control structure, containing a pointer to the work queue, and spawn
the thread, passing a pointer to the control structure.  In the thread
function, loop while there are work items left to process.  Access to the work
queue is synchronized by a function which uses an atomic index into the work
queue.  Once there are no namespaces left to process, return from the thread
function.  Once all threads return, clean up and return.

Bugs Affected: 804754
Bug Action: UPDATE
Bug Comment:
Port for Cray-b2_7
Jira: LU-3970
Gerrit: http://review.whamcloud.com/#/c/7783/
Patchset: PS7/PS8

svn-id: lustre:21023