Store kthread_run return value as task_struct * in ldlm_drop_caches (Cray-b2_7)
In ldlm_drop_caches, store the return value of kthread_run in a variable with a
task_struct pointer type instead of an integer type.  Use the macros IS_ERR and
PTR_ERR to detect and report errors on this return value.
Bugs Affected: 811124
Bug Action: UPDATE
Bug Comment:
Port to Cray-b2_7

svn-id: lustre:21024