Cleanup rpcs from ldlm_drop_caches that are on import replay lists
Job managers, including ALPS, attempt to free memory allocated by Lustre on
behalf of a job when the job terminates. One of the steps in the cleanup is
a call to ldlm_drop_caches(), which cancels the unused ldlm_locks for each
ldlm_namespace. For some applications, dropping locks can generate a large
number of rpcs. An rpc is not freed as soon as it is handled. Instead it is
placed on the import replay list in case it is needed for recovery. The rpc
is not removed from the replay list until the client knows the server has
committed the transaction. ldlm_drop_caches() only waits for the LDLM_CANCEL
rpc to complete; it does not wait for the rpc transaction to be removed
from the import replay list. The result is that a substantial amount of
memory can still be allocated to rpcs when ldlm_drop_caches finishes. In
some cases, enough memory is still allocated that the next job fails.

The solution is to have ldlm_drop_caches() ping each namespace after it
finishes dropping locks. The ping response will contain the id of the last
committed transaction, which should be greater than the transaction id of
the last LDLM_CANCEL rpc. When after_reply() handles the ping response, it
will call ptlrpc_free_committed(), which will clean up the import replay
list.
Bugs Affected: 823598
Bug Action: UPDATE
Bug Comment:
Send a ping to each import at the end of job cleanup. The ping
will trigger cleanup of the import replay list.
Test Description:
Cray-b2.5 version tested.
Developer Notes:
This patch is expected to be temporary. The ping should included in the new
pre/post job functions that are currently being designed as part of the
pingless client project. No plans at this time to push the patch upstream.

svn-id: lustre:21025