LU-5919 build: Make "make rpms" use "make srpm" output
We rework the rpms and srpm Make targets.  The srpm
target generates the source rpm package much as it did
before, but it puts the results in a known temporary
rpm build tree.  This allows the "make rpms" target to
simply depend on the srpm target, and then build the
binary rpms from the source rpm.

Previously the rpm target _also_ generated a source rpm
package, which could be redudant, and failed to verify that
the source rpm package was itself buildable.

This patch is back-ported from the following one:
Lustre-commit: a71ab04971382c77f708d6a082e8ee2fa845e9be
Lustre-change: http://review.whamcloud.com/12961

Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>
Change-Id: Icc493923918f6288827bf8a06bae0d2b1e0a5518
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>
Reviewed-on: http://review.whamcloud.com/16619
Tested-by: Jenkins
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>