MRP-2791 test: test_85a: unused locks error
When we changed FSNAME from lustre to lustre-1, $3
will give “addr” value “clilov” instead of getting
proper address, due to this we are getting "Found no match" error.
But if we change $3 to $NF getting addr value correctly.
Generally NF will give total number of fields in the current
input, but in our case it happens to be last field.

Signed-off-by: Chennaiah Palla <chennaiah.palla@seagate.com>
Change-Id: I7ea8ff97ad449c1a9f06a242f921f41e4b1a9a5c
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/8026
Tested-by: Jenkins
Reviewed-by: Vinayakswami Hariharmath <vinayakswami.hariharmath@seagate.com>
Reviewed-by: Rahul Deshmukh <rahul.deshmukh@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>