Add init scripts to Pinehurst lustre rpm
We would like to use the /etc/init.d/lnet script for bringing up lnet on
Pinehurst nodes when the nodes boot up.  The file was not being packaged
correctly due to its removal in the install section of the spec file, a
non-robust check on whether we are building for CentOS in the autoconf macros,
and the lack of its listing in the files list.  This mod addresses those issues,
but only for the pinehurst builds for now.  We may want to do something similar
for the other builds.
Test Description:
Built pinehurst rpm and verified that /etc/init.d/ scripts were present.  At
this point, it is not important to test whether the other builds work because
we are on a PH-specific branch.
NA

svn-id: lustre:21131