LU-7261 ldiskfs: fix large_xattr overwrite
Handle the case where a large (external inode) xattr is being replaced
correctly.  The special case for this in ext4_set_xattr() was
incorrectly setting the offset of the xattr data within the inode
when it shouldn't have.

Add an e2fsck check of the large_xattr filesystem in conf-sanity
test_61 to verify this is working correctly.

This patch is back-ported from the following one:
Lustre-commit: 5155ce86143c6937800b2883c1c433ee73be884a
Lustre-change: http://review.whamcloud.com/16777

Signed-off-by: Alexey Lyashkov <alexey.lyashkov@seagate.com>
Change-Id: I14fd2f4ca8491a10af732a6aca36dcddd410374f
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-on: http://review.whamcloud.com/16881
Tested-by: Jenkins
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>