LU-7543 obd: reserve connection flag OBD_CONNECT_LOCK_AHEAD
The connection flag OBD_CONNECT_LOCK_AHEAD will be used for the
following the patch: LU-6917 LDLM lock ahead
http://review.whamcloud.com/13564

Note that the original change to master was mistitled "LU-7543 obd:
reserve connection flag OBD_CONNECT_OBDOPACK".

This patch is ported from the following one:
Lustre-commit: c801f497807f45e7e081f0e1ea11037749a60270
Lustre-change: http://review.whamcloud.com/#/c/17645/

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I75171f5e9095d40481b4462de74c6a63e1cf3084
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Patrick Farrell <paf@cray.com>
Reviewed-on: http://review.whamcloud.com/18413
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>

Notes on port to feature-2.7:
The only thing we needed here was the addition of "lock_ahead" to the
obd_connect_names array in lustre/obdclass/lprocfs_status.c. However,
I chose to adopt their locations for everything else in the patch with
the goal of reducing the likelihood of future conflicts with
b_neo_stable.