LU-7543 obd: reserve connection flag OBD_CONNECT_FLAGS2
This is a feature for the client and server to use
obd_connect_flags2 to communicate future feature flags. The
client should set this flag whenever any flags in that field
are requested, and the server should mask unsupported features
from this field (assuming it understands OBD_CONNECT_FLAGS2).
When checking if an OBD_CONNECT2_xxxx feature is supported,
the client/server needs to firstly check if OBD_CONNECT_FLAGS2
is supported, since this field is also beyond the end of the
old obd_connect_data.

This patch is ported from the following one:
Lustre-commit: a5dc6abb06215856dde2d572f7d4c24dddebb718
Lustre-change: http://review.whamcloud.com/17647

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I466a4f858326d79f8c732116915b078902aa0736
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Patrick Farrell <paf@cray.com>
Reviewed-on: http://review.whamcloud.com/18415
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>