Renamed Makefile to Makefile.obs
The Makefile is overwritten during the configuration process, and git
is tracking it, so it always appears in the list of changes, which
gets quickly annoying when working on the sources.

So rename the Makefile to Makefile.obs to get it out of the way. Fix the
OBS config so it doesn't package it either.

Whatever is using that Makefile can use it with:

    make -f Makefile.obs