LU-7042 lnet: Handle OFED 3.18 packaging definitions
Starting with OFED 3.18 the OFED package started to use config.h
provided by autoconf tool. This lead to clash between PACKAGE_*
macros which are defined in OFED and Lustre headers.

Also dealing with kernels that have lustre enabled already
required to undefine the common macros that used by in-kernel
Lustre client and this sources.

This fix undefine all symbols that are generated by Lustre autoconf
to avoid conflicts with kernel defines or OFEDs. They are undefined
right before new definition in config.h. The list of symbols to undef
is automatically generated by autoconf and should not be extended in
the future.

Also undefine clashed macros in autoconf checks.

Lustre-Commit: bfa4bc911208f7a7a447984a40266c70ad8814ab
Lustre-Review: http://review.whamcloud.com/16418

Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>
Change-Id: I77fbfd5f6add78e63870f84da7928b74ac48cadd
Reviewed-on: http://review.whamcloud.com/17034
Tested-by: Jenkins
Reviewed-by: Minh Diep <minh.diep@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>