Add module parameter reg_fail_timeout to allow continuous registration failures to kill node.
During network outages on very large machines, it is possible to use up
all of GART space with connections that are in purgatory waiting to be
freed when we finally make a new connection.
This mod adds a timeout parameter so that when we fail registering
memory for fma blocks for a period of time, we can bring the node down
so it is not stuck in a state of being up but unusable.
This can only happen on service nodes as there can potentially be 10s
of thousands of connections.
A recommended setting for reg_fail_timeout would be 60 - 300 seconds.
The default setting for reg_fail_timeout is -1 (disabled).

Bugs Affected: 821876
Bug Action: RESOLVE
Test Description:
Set fail_loc 0xf002 which fails memory registrations and see that we
BUG after the required timeout.
Test that transient registration failures within the timeout period
do not cause BUG.

svn-id: lustre:21308