Pass configure_args to rpmbuild when building an rpm source
When using make rpms, which builds the rpms with lustre.spec,
configure_args was not set. So the kdir and kobjdir in the specfile
would always use the defaults. When the --with-kernel or
--with-kernel-obj was passed on the configuration command line
pointing to a different kernel, the rpms source name would contain the
version of the running kernel, not the desired kernel.

The binary rpms were not affected.