MRP-3063 test: Deregister changelog client on test fail.
The tests which are related to changelog did not deregister
the changelog client when any failure was encountered,
because of which any changes afterward were also recorded
leading to huge test stdouts. I have added the functions
error_and_deregister which will first deregister the
changelog client created by the test before exiting.

Signed-off-by: Kirtankumar Krishna Shetty <kirtan.shetty@seagate.com>
Change-Id: Idbf31f7eff5725a79635e9c7db1cb68760e69732
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/9079
Tested-by: Jenkins
Reviewed-by: Vladimir Saveliev <vladimir.saveliev@seagate.com>
Reviewed-by: Alexander Lezhoev <alexander.lezhoev@seagate.com>
Tested-by: Parinay Vijayprakash Kondekar <parinay.kondekar@seagate.com>