LU-4840 lfs: Use file lease to implement migration
This is a port of master commit:
85bd36cc69563d7a79e3ed34f8fadb4ed1a72b7c
http://review.whamcloud.com/10013]

To Cray b2_7_fe.

The main difference was our 2.7 still contains the size on MDT code
which is gone in master.  Since that code is already inactive in 2.7,
we can mostly ignore it and do not need the removal patches.

Original description:
Implement non-blocking migration based on exclusive open instead of
group lock. Implemented exclusive close operation to atomically put
a lease, swap two layouts and close a file. This allows race-free
migrations.

Make the caller responsible for retrying on failure (EBUSY, EAGAIN)
in non-blocking mode.

In blocking mode, allow applications to trigger layout swaps using a
grouplock they already own, to prevent race conditions between the
actual data copy and the layout swap. Updated lfs accordingly. File
leases are also taken in blocking mode, so that lfs migrate can issue
a warning if an application attempts to open a file that is being
migrated and gets blocked.

Timestamps (atime/mtime) are set from userland, after the layout swap
is performed, to prevent conflicts with the grouplock.

lli_trunc_sem is taken/released in the vvp_io layer, under the DLM
lock. This re-ordering fixes the original issue between truncate and
migrate.