LDEV-303 snapshot: avoid destroying MGS snapshot piece
When destroy the snapshot, to improve the performance, we destroy
the snapshot pieces on every server (MGS/MDT/OST) in paralle. But
if the snapshot is mounted with original independent MGS running,
and if not specify the "-f" option, then it will cause only the
snapshot piece for MGS to be destroyed, and then the subsequent
snapshot_destroy with "-f" option cannot find the snapshot piece
for MGS.

This patch adds wait process between destroying snapshot piece(s)
for OST and the one(s) for MDT/MGS. That avoids such trouble and
will NOT affect the performance too much. It is some compromise.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ibfdffe642cb725e6a35a7b783ca5dbb83e6b59cf
Reviewed-on: http://review.whamcloud.com/18363
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Tested-by: James Nunez <james.a.nunez@intel.com>
Reviewed-by: Jinshan Xiong <jinshan.xiong@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>