LDEV-302 tests: check snapshot status properly
If the MGS and the MDT0 are on separated devices, then when mount
the snapshot, the snapshot piece for the MGS will not be mounted
if the current MGS is still running. Under such case, the command
"lctl snapshot_list" will show the MGS piece status as 'not mount'.
That is normal. The test scripts should handle such case properly.

On the other hand, when umount snapshot, if some snapshot piece(s)
is not mounted, then ignore it, not report failure.

Test-Parameters: mdtfilesystemtype=zfs mdsfilesystemtype=zfs \
ostfilesystemtype=zfs mdtcount=1 testlist=sanity-lsnapshot

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I3a81fb85896a66fe682c1464b01baa4d7d2185b2
Reviewed-on: http://review.whamcloud.com/18359
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>