LU-6866 hsm: prevent migration of HSM archived files
The reference copytool cannot handle migration of HSM archive
files. In the MDT migration path check for HSM attributes and fail if
they are present. In the LMV layer allow creation of volatile files
with any MDT index. Add a test to sanity-hsm to ensure that attempting
to migrate an HSM archive file is handled safely.

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Change-Id: I559609d439c8f1fcf3e80770d1e3cc8bd59713fa
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Sergey Cheremencev <c17829@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>
Reviewed-on: http://es-gerrit.dev.cray.com:8080/#/c/17338/
Tested-by: Jenkins