eorganize changelog clear/user remove code
Originally this was done all in one function, split up the function
according to what each piece does.
Remove spawning a thread, as it prevents the end user from knowing
the result.
Check that the changelog user exists, and that the range is within
the changelog.  Add more comments and rename structures.

Bugs Affected: 828641
Bug Action: UPDATE

Signed-off-by: Ben Evans <jevans@cray.com>