LU-7604 gss: handle revoked keys properly
Revoked keys are no longer returned by request_key kernel
function. So it is now necessary to remove revoked keys from
keyring when flushing context.
Moreover, if a revoked key is present, do not consider it
matches when searching for a valid key with request_key. That
way it will be replaced with a valid, newly created one.

This patch is back-ported from the following one:
Lustre-commit: fa4646291c7059597bd5d682f0c6c89b65ecf97e
Lustre-change: http://review.whamcloud.com/17721

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I0fbaf01f6a8f50e9fb06eef96c74e73c25de257f
Reviewed-by: Jeremy Filizetti <jeremy.filizetti@gmail.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Reviewed-on: http://review.whamcloud.com/18004
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>