LU-6179 llite: Implement lock ahead
Lock ahead is a new feature allowing userspace to request
extent locks in advance of the IO which will use them.
These locks are not expanded beyond the size requested by
userspace.  Their use is intended to address lock
contention between multiple clients resulting from lock
expansion.  They should allow optimizing various IO
patterns.  (Further information in LU-6179)

Asynchronous glimpse locks are a speculative version of
glimpse locks, and already implement the required behavior.
Therefore, this patch takes the CEF_AGL flag and renames it
to CEF_SPECULATIVE.

There is no change to the behavior the flag invokes, but it
is now shared between AGL and lock ahead requests.

Additionally, lock ahead creates extent locks in advance of
IO, and so breaks the assumption that the holder of the
highest lock knows the current file size.

So we also modify the ofd_intent_policy code to glimpse
every PW lock and merge their LVBs.

Currently, lock ahead lock requests are always
non-blocking.  The hope is to change this in a future
version, but for now it avoids a number of deadlocks.

LDLM_FL_BLOCK_NOWAIT was only implemented for group locks,
so this patch also makes the changes necessary to allow it
for standard extent locks.

Finally, this patch also adds request only locking as
a per-file flag which can be set with an ioctl.  This
allows user space to control whether or not lock requests
on this file will undergo lock expansion.

This means if lock ahead locks are not created ahead of IO
(due to inherent raciness) or are cancelled by a competing
IO request, the IO requests that should have used the
lock ahead locks will not result in expanded locks.
This avoids lock ping-pong, and because the resulting locks
will not extend to the end of the file, future lock ahead
requests can be granted.  Effectively, this means that if
lock ahead is interrupted by a competing request, it can
re-assert itself.

This is the 2.7 version of this code.  The master version
has seen extensive testing, the 2.7 version has not yet.
This pull request is to get the review process started.

Bugs Affected: 825944
Bug Action: UPDATE