MRP-2586 test: Wait for archive operation to be completed
As bandwith control is set to 1048576B/s in the test, so
the archive operation takes <= 40 secs to copy 39000000
bytes.

If we dont wait for the on going archive action to finish
then status will remain as STARTED and never becomes
SUCCEED as we call copytool_cleanup soon after archiving
5MB of data and the request will be stuck there for ever.

Sometimes because of previous unclean shutdown of copytool,
test_60 results in failure because of stuck request.

Signed-off-by: vinayakswami hariharmath <vinayakswami.hariharmath@seagate.com>
Change-Id: I0d84329e47c700417410dcdd0ad0029427da12ef
Signed-off-by: Kirtankumar Krishna Shetty <kirtan.shetty@seagate.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/8671
Tested-by: Jenkins
Reviewed-by: Ujjwal Lanjewar <ujjwal.lanjewar@seagate.com>
Reviewed-by: Alexander Lezhoev <alexander.lezhoev@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>