LU-7353 utils: fix lctl usage messages
Fix the lctl usage message for sub-commands that need the LNet network
to be specified.  lctl::g_net_is_set() incorrectly recommended using
the "network" command to specify the LNet network, which is incorrect
when using lctl in command-line mode:

  # lctl peer_list
  You must run the 'network' command before 'peer_list'.
  # lctl network tcp0 peer_list
  # lctl --net tcp0 peer_list
  12345-192.168.20.1@tcp [1]192.168.40.147->mookie-gig:988 #15

Fix that to correctly recommend using the "--net" command when using
non-interactive mode, and to return an error if "network" is used in
non-interactive mode with extra arguments.

Replace mention of "portals" in help messages with "LNet".
Remove mention of obsolete elan, qsw, ra network types.

Improve the help message content for related subcommands.
Fix whitespace and command descriptions for related subcommands.

This patch is back-ported from the following one:
Lustre-commit: fa54edb5af34c3bcbf878d7498ef337ef3a1df7a
Lustre-change: http://review.whamcloud.com/16980

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Ib092b146e492f080eb74a17d6b8a495b4a77f24a
Reviewed-on: http://review.whamcloud.com/17866
Tested-by: Jenkins
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>