bugfix 834643
A quick and dirty fix for bug 834643. ko2iblnd-probe is not being
installed in the correct place so a symlink is now created to link
the file to the correct place. This is not a good long term solution
but the builds will need to be refactored for a good long term solution
to be implimented. I installed the rpm in an image root and verified
that it created the symlink as expected. I also verified that the
symlink was removed when the package was uninstalled.

Bugs Affected: 834643
Bug Action: RESOLVE