Build the -devel package directly from lustre
First start by folding the cray-lnet package into the main build
so that we can package all the needed file into the -devel rpm
in one go.

The main goal here is to package the Module.symvers file so that
the DVS build can find the required symbols.

This should allow us to turn on the kernel's modulesymvers option
finally.

Bugs Affected: 808203
Bug Action: RESOLVE