LU-7173 mdt: intent vs unlink race
a race between intent and unlink results in working on stale object,
the intent resend finds the lock and !existent object and LBUG:
    ASSERTION( ((o)->lo_header->loh_attr & LOHA_EXISTS) != 0 )
in mdt_getattr_name_lock()->lu_object_attr(). a check for getattr is
added for !existent object in RESENT case.

another case is when resend finds inode cached on thread,
osd_iget_check() checking the inode returns ESTALE as nlinks==0.
ldlm_lock_enqueue() gets an error from ns_policy() and tries to
destroy the found lock, which is granted, getting "lock still on
resource" LBUG.

It is unclear if the lock reached the client in the original reply,
just leave the lock on server, not returning it again to the client.
Due to LU-6529, the server will not OOM in case the 2nd reply will
be handled on the client.

This patch is back-ported from the following one:
Lustre-commit: 9b520c36ddff687c51bea026d78380eb85981971
Lustre-change: http://review.whamcloud.com/17501

Seagate-bug-id: MRP-3042
Signed-off-by: Vitaly Fertman <vitaly.fertman@seagate.com>
Change-Id: I8ca2f8c4f58f9bbbef35e55418f93c1e6bcb28bb
Reviewed-on: http://review.whamcloud.com/18075
Tested-by: Jenkins
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Bobi Jam <bobijam@hotmail.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>