Consolidate lustre specs to one spec
I consolidated the lustre spec files into one spec file in order
to improve supportability and usability from the standpoint of obs.
The build now uses the upstream spec file. The cray-lustre.spec
runs autogen, configure, and make on the upstream spec file in the
obs build root. This allows us to build our builds from the upstream
spec file without having to make changes to it ourselves.

The new build projects are as follows:

    cray-lustre:Cray-b2_7_fe:CLFS cray-lustre
    cray-lustre:Cray-b2_7_fe:XC:compute cray-lustre
    cray-lustre:Cray-b2_7_fe:XC:dal cray-lustre
    cray-lustre:Cray-b2_7_fe:XC:server cray-lustre
    cray-lustre:Cray-b2_7_fe:XE:compute cray-lustre
    cray-lustre:Cray-b2_7_fe:XE:dal cray-lustre
    cray-lustre:Cray-b2_7_fe:XE:server cray-lustre
    cray-lustre:Cray-b2_7_fe:elogin cray-lustre

The names of the build projects indicate the product which the builds
are to be run on. To determine the available OS's for each build
project run, for example:

$osc pr cray-lustre:Cray-b2_7_fe:XE:compute

To build the new packages run, for example:

$obs build cray-lustre:Cray-b2_7_fe:CLFS cray-lustre --repo CENTOS_6_5_DMP

I tested all the builds and I installed the packages on saturn-p2
and ran a quick test to see if lustre worked. The tests ran
succesfully.

Bugs Affected: 801143
Bug Action:  RESOLVE