LU-7828 statahead: set sai_index_wait with lli_sa_lock held
It is the sponsor thread of the statahead thread to update the
sai::sai_index_wait. Originally, it didn't hold the lli_sa_lock
when did that. Becuase of out-of-order execution others may miss
to wakeup such thread.

On the other hand, if the statahead RPC gets failure, it should
wakeup the sponsor thread, not the statahead thread.

Lustre-commit: 35fb132546ff45cd4b3f08dd9feaa6f69ad02b3a
Lustre-change: http://review.whamcloud.com/18499

Signed-off-by: Li Xi <lixi@ddn.com>
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ib14e52a51b6eb1bfcea534ccfe6476d21d7953ce
Reviewed-on: http://review.whamcloud.com/18681
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Lai Siyao <lai.siyao@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>