Lock ahead: Stop glimpsing at first active lock
Previously, lock ahead would glimpse every PW lock until
it reached the previously known file size.  This means an
impact on the server side whether or not lock ahead is in
use.

With this change, we now only glimpse until we find a lock
we know is 'active'.  (We look specifically for non
request-only locks, since lock ahead locks are always
request only.)

This is explained in more detail in comments in ofd_dlm.c.