MRP-2960 ldlm: Implement lock ahead
This is the server side of the lock ahead feature.
The client side of this feature is not compatible with
Lustre 2.5.

The server side patch has the following components:
New LDLM flag, LDLM_FL_NO_EXPANSION, telling the server not
to expand extent locks requests with this flag set
Changes to lock matching code to allow non-blocking lock
requests for all lock types (not just group locks)
Changes to the ofd intent code to get correct file sizes
with locks in advance of IO

The description of the full feature (including client code)
is here:
http://review.whamcloud.com/#/c/13564/

Signed-off-by: Patrick Farrell <paf@cray.com>
Signed-off-by: Vitaly Fertman <vitaly.fertman@seagate.com>
Change-Id: Iba6fd5fc3190bab75e5d9a6b7bc48e267a32341b
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/9909
Tested-by: Jenkins
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>